Update Printing.CUPS* histograms
Extend the times on most of our histograms with upcoming expirations
because they are still useful for monitoring and investigations.  Also
replace skau with bmgordon as the owner, since skau has left the team.

The Printing.CUPS.HighestIppVersion metric is ready to be deprecated, so
that one is left alone to be removed separately.

Bug: 1303776
Bug: 1306158
Bug: 1303779
Bug: 1303778
Change-Id: Iae4912ad325a822ad32be6e2415bc926bb1d849c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3523422
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#980876}