breadcrumbs: pass weak ptr to write timer
BreadcrumbPersistentStorageManager has a OneShotTimer member that it
uses to delay writes to file. A OneShotTimer takes a callback and runs
it after a given delay.

Currently, a raw pointer to `this` is passed to the OneShotTimer,
which can cause a crash if the object is destroyed before the timer's
callback runs. This change makes it a WeakPtr, changing the crash to
a no-op.

Bug: 1305954
Change-Id: Ifafdf9754a2f2ff55836c557947305ecad0a7f4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3522792
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/main@{#980895}