Add player to handle looping logic in second animation theme.
Up until now, the logic was really simple: Play the animation in a loop
starting at time 0 (the first frame) after each animation cycle ends.
This is not always the case anymore. The second animation theme will
have a marker in it, which denotes the timestamp at which the animation
should be restarted after each cycle ends. The marker can be any
timestamp in the middle of the animation cycle. As such, a new class
has been added to manage this logic.

I have intentionally not added the unit tests in this CL for practical
reasons: My solution to unit testing involves some changes in the
compositor, which can be controversial. We want this to land by M101,
so I don't want to risk this CL being blocked by those changes and not
making it. Thus, I'll submit this code first, then put the CL for unit
tests immediately up afterwards (and that CL can potentially miss M101).

Bug: b:223672841
Change-Id: I503a5aa65734ffc72d46a95de99f4f3db0f7bb92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3523442
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#980905}