NativeWidgetNSWindowFullscreenController: Add cross-screen fullscreen
Re-add NativeWidgetNSWindowFullscreenController, this time disabled
by default.

When an NSWindow is told to go fullscreen on a display that it is
currently not on, use -[NSWindow setFrame:display:animate] to move
the window to the target display.

Set the parameter animate:YES, because it looks nicer. As a happy
coincidence, this allows us to immediately call -[NSWindow
toggleFullscreen:], and the fullscreen transition appears to be"queued"
behind the move to the new display. Even happier is that this appears
to side-step all of the issues we had previously seen involving loss
of focus. Leave a note referencing the workaround for the the loss of
focus issue, in case it reappears.

Similarly, when the NSWindow exits fullscreen, again use setFrame
again with animate:YES to animate the window back to its original
position.

Update the NativeWidgetNSWindowFullscreenController::State state
machine to include the new states kWindowedMovingToFullscreenTarget
and kWindowedRestoringOriginalFrame, for these new states.

Force communication from NativeWidgetNSWindowFullscreenController to
NativeWidgetNSWindowHost and its NSWindow to go through a Client
interface.

Using the new interfaces, add lots of tests for various edge cases
(e.g, being told to transition to fullscreen on a specific display,
while in the middle of a transition to fullscreen on another display).

Bug: 1302857, 1210548
Change-Id: Ie7bf034a93aadb0f359d73a14bc2487f74f2c946
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3519784
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#980947}