Use realpath instead of abspath in native_hosts/echo.py
On MacOS X /var is a symlink for the /private/var. Use realpath
instead of abspath for comparison, because cwd returns path with
expanded symlinks.

On Windows cwd returns path with user name in 8.3 format.
For user name "longusername" it returns a path that starts with
"c:\users\longus~1\appdata\local\temp\". Use realpath as well,
so full user name is used in path.

Change-Id: I3e61dc7230d4e19688b1959be3b84719bc14d8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3516272
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Tomasz Moniuszko <tmoniuszko@opera.com>
Cr-Commit-Position: refs/heads/main@{#981033}