Add underline and footer for settings textarea
This CL fixes the underline which should be blue when focused. It also
makes the underline red when the invalid attribute is set.
The CL also introduces a footer below the underline. By default the
footer is gray. When invalid, the footer will be red. If there are no
footers supplied as a property, no footer will be shown.

Further CLs will add this functionality to password edit dialog note
section. Current implementations are not affected.

Mocks
- Focused: https://screenshot.googleplex.com/8oicWiMX95whMXa
- Invalid: https://screenshot.googleplex.com/7GrajkKoaxjAUzN
Implementation
- Focused: https://screenshot.googleplex.com/8HSMH4qjVoRM5z5
- Invalid: https://screenshot.googleplex.com/BKQXXUok2HCQ9UD

Bug: 1298027
Change-Id: Ida83b6df8ceafae8f4283681a6417b2a3d3f7478
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3515733
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Adem Derinel <derinel@google.com>
Cr-Commit-Position: refs/heads/main@{#981034}