Consider previous state for running screen shares
Due to the current implementation of DlpContentManager, we can call
CheckRunningScreenShares() multiple times for essentially the same user
action (e.g. one click on a tab). This can cause multiple warn dialogs
on top of each other, as well as duplicated report events.

With this change, we store the previous restriction and the content
that caused it in the existing ScreenShareInfo structure, so that we
only react to changes in any of these thus eliminating duplication.

Fixed: 1300565
Change-Id: I59829dcda4ade4bdbd4aaefcd90748ed810c4e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3510407
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981064}