Privacy Sandbox Settings: Prevent consent being closed without decision
CL subclasses the views::DialogDelegate and overrides the
OnCloseRequested function to prevent the dialog from being closed, if it
is a consent, by any action other than an explicit Close() call on the
widget itself.

As the only way to call Close() is to make a decision in the consent,
this prevents it from being shut without a decision having been made.

Bug: 1286384
Change-Id: I3ef514ac0e8d0af393a72959343c67ab660bf056
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3524406
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#981067}