Privacy Sandbox Settings: Do not show dialog for users who block 3PC
CL adjusts the behavior of the GetRequiredDialogType() function to
prevent users who block 3PC from receiving the dialog. This is sticky,
and once a decision was made to not show the dialog, it will not be
shown again.

CL also adjusts some of the check ordering, so that changes in the
profile do not retroactively change why a dialog was shown.

Bug: 1286276
Change-Id: Ib502ff4c5549f56113cac50f9458caf0e47bcec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3515978
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#981068}