Added fallback strings for an upcoming onboarding experiment.
UMA data shows that ~12% of all users do not see the correct strings due
to a timeout when fetching them. As a short-time workaround, this CL
adds fallback strings to be used in case of such timeout. A proper
solution is being discussed in the bug linked below.

Bug: b/223553250
Change-Id: I3faa512bad1b160eb86ca5768904265614bdffc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3515933
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Auto-Submit: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/main@{#981070}