[Autofill] Fix cache-downdating problem.
This CL introduces a workaround for the cache-downdating issue.
This problem arises when AutofillManager::GetCachedFormAndField()
is called with an outdated FormData.

Generally, this can happen when a FormData is stored as member
variable which is then passed to GetCachedFormAndField().

In particular, when the user focuses a field, the current FormData
is stored in an AutofillExternalDelegate member. While the Autofill
popup is shown, the cache may receive further updates. When the
user selects a suggestion, AutofillExternalDelegate passes the
now-outdated member variable to FillOrPreviewForm(), which calls
GetCachedFormAndField(), which replaces the cached form with
the outdated form.

This is especially common with cross-frame forms: fields were
invisible at page load, and when a field is focused, the fields
from the other frames are re-extracted, which causes cache updates.
However, these cache updates happen while the Autofill popup is
being displayed. When the user selects a suggestion, these cache
updates are reverted.

To work around this problem, this CL introduces a FormVersion member
of FormData that monotonically increases whenever a FormData is passed
from ContentAutofillDriver to AutofillManager. The cache replaces an
entry only if the new candidate is at least as fresh as the existing
entry.

An alternative solution would be to not store FormData objects in
member variables.

Bug: 1117028, 1187842
Change-Id: I15c1df448b25e42348eea084fe3efd39207c4fd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3522392
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#981072}