Remove configurable report-allowed callback from attribution manager
This was added in crrev.com/c/3421416 for the API simulator, but it is
unnecessary because the default behavior of the method that the real
manager uses allows all operations, which is exactly what we want for
the simulator:
https://source.chromium.org/chromium/chromium/src/+/main:content/public/browser/content_browser_client.cc;l=472;drc=4d2b5adb556128aab9313fc3851f192c254e09cb.

This change will make it easier to address crbug.com/1305577.

Additionally, we will eventually want a non-null storage partition in
the simulator to support aggregatable reports there, so this begins that
process as well.

Bug: 1305577
Change-Id: I67d0b7bbae33e2464898e0e6e38d67e8beb5ab45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3517336
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981089}