Remove rmcilroy@chromium.org from third_party/blink/web_tests/external/wpt/requestidlecallback/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I0d96e725215b046cb3afd23ffd730757351e1dca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501583
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981103}