[FRE] Disable the signin/out buttons when signin/out is in progress in signin FRE
This cl adds a property field to register when a signin or signout event
starts in SigninFirstRunProperties. While signin or signout is in
progress the "Continue as ..." and "Continue without an account" buttons
are disabled. In addition the selected account row is also disabled.

The property field will allow a loading spinner to be added to the UI
in future cls.

Bug: 1294994
Change-Id: I2e712389e9b0ce69786d9e211c9843a3da997b86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3474376
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981104}