Remove pinpoint_traces.py
The pinpoint_traces.py tool uses the deprecated pinpoint_cli tool. It is
unclear how much is required to make it work with the new 'pinpoint'
tool from depot_tools.

Remove the tool to avoid confusion.

I just discovered these flags that may be useful for smb to restore the
functionality in the future:

> pinpoint get-job -name=... -download-artifacts -work-dir=...

Bug: 1166441
Change-Id: Iacbcb853a7ab51d6174cac8dd48db68af9b7eb63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3497462
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981108}