personalization: Add Google Photos API call to fetch enterprise setting
This CL introduces and implements a new mojom API call for the Wallpaper
app to fetch the user's Google Photos enterprise setting. Possible
setting values are "enabled" (the user is not a Dasher user or is
explicitly permitted to access Google Photos), disabled (admin has
disallowed Google Photos for this Dasher user), or error (the setting
could not be fetched or was unknown on the server side).

This setting will be used by the UI to indicate when a user cannot
access the Google Photos wallpaper tile due to enterprise policy.

Fixed: b/223657704
Change-Id: I938bc53fef5b2fd006a29d3d1d7e64a71559c091
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3514896
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Colin Kincaid <ckincaid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981136}