Fix network screen skip when welcome screen is also skipped
When the welcome screen is skipped (e.g. during Chromad migration -
crbug.com/1209246) the network screen can be shown before the device
connects to a network. This race condition was causing the skip on
Ethernet connection to faiil.

This CL modifies the network screen to also check for an Ethernet
connection as soon as the device gets connected to a network. This
covers the corner case described above.

Bug: 1256890
Change-Id: Ie0ea45eb74d9a9ca5425649608603d90e66107c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3512624
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>
Reviewed-by: Roman Aleksandrov <raleksandrov@google.com>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981166}