[infra] Unwrap singular ignore_defaults in args.listify.
Previously, when calling listify with an ignore_default object as an
argument, its value was not flattened if necessary before being
unwrapped. This meant that subsequent calls to listify with the result
of a listify call that took an ignore_default object would result in
more deeply-nested lists. This change performs the ignore_default
unwrapping for each argument so that it will properly be flattened into
the result.

Change-Id: I0580497e1a1f15bedbb82871adabf3ec6972aa16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3525977
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981204}