Move full themes to /browser_ui/theme/.
This creates a new /browser_ui/ package that is allowed to depend on
both /widget/ and /styles/. Often there were activity/app level themes
that needs both of these dependencies, but clearly were not widgets.
There shouldn't be a need for /widget/ to ever depend on this new
package.

Bug: 1303856
Change-Id: Iff427dd8ba4585a874a129ad9cb46d14be5e8cfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3518425
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981208}