Use a focus ring color that draws attention for help bubble close button
Previously, the close button used a focus ring that was hard to see and
did not meet minimum accessibility requirements. This led to it not
being clear when a bubble was focused if the close button was the active
view (even if the user was not low-vision!)

The new ring is just over 1dip thick and white (similar to the other
focus rings used in the bubble for buttons). It is very obvious when it
is focused.

Change-Id: Ia0a1aa16bfaaba803c83025bc6182e7ea42f2ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3524388
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981228}