Center the separator between recent apps and app grid
This CL centers the separator to be spaced equidistant from the bottom
of the recent apps and the top of the first apps grid row.

If no recent apps are shown or if a sort nudge/toast is shown, then the
separator is placed directly below the continue/recent apps section with
no centering.

The separator is moved from |continue_container_| to be a child of
|scrollable_container_|.

Bug: 1288680
Change-Id: Iafdc715ce3adb6798dfad42881013e82b745105e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3510776
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981296}