[side search] Update tooltip and accessibility strings
This CL updates the tooltip and accessibility strings for the side
search toolbar buttons and the close button. Both accessibility and
tooltip strings now take into account the active state of the button.

This CL also addresses a small bug with the tooltip not being
updated correctly for the toolbar button and eliminates the
SetLabel() call for the SideSearchIconView which would cause a full
text presentation in some cases in the omnibox.

Bug: 1306372
Change-Id: I4c2225f77568bda2184054d9be4c165e9fb723a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3524270
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981301}