Don't avoid disambiguation dialog in CCT/WebLayer if Chrome can't handle
For external protocols, if we block the outgoing intent because it'll
show the disambiguation dialog, we end up landing on a page that just
says '"<protocol>" links are blocked'.

Given the goal of blocking the intent chooser is to avoid showing the
Chrome icon showing for navigations the CCT/WebLayer could handle, it
should be fine to show the dialog for things that are explicitly
supposed to leave the CCT/WebLayer (through the use of external
protocols).

Bug: 1305256
Change-Id: Icaf6b9e5058f35e89fb191e0e7b9b3b92ce67393
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3517311
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981353}