[lacros] Introduced a method for creating the browser from template.
This is the fourth part of changes that will implement the functionality
explained in the linked crbug.  The patch is based on earlier work by
mkim@igalia.com [1].

This patch introduces a Mojo method for creating a Lacros browser with
the tab strip saved earlier as the template.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3341786

Bug: 1268741
Change-Id: I875825e1a6c8c77f573e7db7660249ba3a5606bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3516257
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/main@{#981359}