Add views::FullscreenControllerMac feature
Control the new remote_cocoa::NativeWidgetNSWindowFullscreenController's
usage by a base::Feature. Leave both the old and new path in place, to
avoid code churn in the event of failing tests.

In NativeWidgetMac::InitNativeWidget pass the parameter
fullscreen_controller_mac_enabled across mojo to determine if the
FullscreenController is to be created. This is done instead of
querying a base::Feature because base::Features are not synced to
app shims.

In BrowserView::ProcessFullscreen, select the path to use by specifying
either a delay or a target display ID to Widget::SetFullscreen. Plumb
the target display ID through from Widget to NativeWidgetPrivate, its
implementation NativeWidgetMac, and finally to
NativeWidgetMacNSWindowHost.

At the NativeWidgetMacNSWindowHost level, select between two different
mojo interfaces (SetFullscreen vs Enter/ExitFullscreen) based on the
feature.

This change should be a no-op, unless FullscreenControllerMac is
enabled.

Bug: 1302857
Change-Id: I6f8f52d0e910347ae4625bf16a5f7fdb97869788
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3522752
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981363}