SET: Handle empty divs for both old and new image contents.
The approach to handling the empty divs is somewhat different for
new and old contents:

For new contents, we use live snapshots so viz isn't aware that the
render pass may be missing (and we'd like to retain the DCHECK on that
side for debugging purposes). So, we let
cc::DocumentTransitionContentLayerImpl know whether it's a live snapshot.
If it is, and the referenced resource_id does not appear anywhere in the
currently generated frame, then we simply append a transparent quad
instead.

For old contents, we send the resource id with an invalid render pass to
viz. When processing the save directive, we pull out these resource ids
as "valid empty resources", and in the future check whether the resource
id is known empty when we're replacing shared element quads.

R=khushalsagar@chromium.org

Change-Id: I00b9990957024b5c00445c36369c6b947a3b950d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3519718
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#981366}